Make greyVest the default theme
