+++ .storybook/preview.js
@@ -3,7 +3,7 @@ import { configure } from 'mobx'
 import greyVest from '../src/themes/greyVest/index.js'
 import Fonts from '../src/greyVest/Fonts.js'
 import Style from '../src/greyVest/Style.js'
-import { ThemeProvider } from '../src/utils/theme.js'
+import { ThemeProvider } from '../src/utils/theme/provider.js'
 
 configure({ enforceActions: 'never', useProxies: 'never' })
