+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.54.16
+
+- Remove unmaintained themes. Make grey-vest the default
+
 # 2.54.15
 
 - Migrate to storybook v7 and fix lots of stories
