+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "contexture-react",
-  "version": "2.54.15",
+  "version": "2.54.16",
   "description": "React components for building contexture interfaces",
   "type": "module",
   "main": "dist/esm/index.js",
@@ -48,9 +48,6 @@
     "@babel/core": "^7.20.12",
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
-    "@date-io/moment": "^1.3.9",
-    "@material-ui/core": "^4.3.3",
-    "@material-ui/pickers": "^3.2.3",
     "@mdx-js/react": "^2.2.1",
     "@storybook/addon-actions": "^7.0.0-beta.27",
     "@storybook/addon-docs": "^7.0.0-beta.27",
@@ -72,7 +69,6 @@
     "eslint-plugin-react": "^7.32.0",
     "eslint-plugin-storybook": "^0.6.10",
     "glob": "^8.0.3",
-    "material-ui-chip-input": "^1.0.0",
     "mobx": ">=4.0.0",
     "mobx-react": ">=6.3.0",
     "mobx-utils": ">=5.0.0",
