--- src/themes/blueberry/Fonts.js
@@ -1,16 +1,0 @@
-import React from 'react'
-
-export default () => (
-  <div>
-    <link
-      href="https://fonts.googleapis.com/css?family=Lato:400,600,700,900"
-      rel="stylesheet"
-    />
-    <link
-      rel="stylesheet"
-      href="https://use.fontawesome.com/releases/v5.3.1/css/all.css"
-      integrity="sha384-mzrmE5qonljUremFsqc01SB46JvROS7bZs3IO2EmfFsd15uHvIt+Y8vEf7N7fWAU"
-      crossOrigin="anonymous"
-    />
-  </div>
-)
