--- src/themes/material/Box.js
@@ -1,6 +1,0 @@
-import { Paper } from '@material-ui/core'
-import { withStyles } from '@material-ui/core/styles/index.js'
-
-export default withStyles({
-  root: { padding: 30 },
-})(Paper)
