--- src/themes/material/Root.js
@@ -1,29 +1,0 @@
-import React from 'react'
-import { CssBaseline } from '@material-ui/core'
-import MomentUtils from '@date-io/moment'
-import { MuiPickersUtilsProvider } from '@material-ui/pickers'
-import Style from './Style.js'
-
-let Root = ({ children }) => (
-  <>
-    {/* fonts */}
-    <link
-      rel="stylesheet"
-      href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap"
-    />
-    <link
-      rel="stylesheet"
-      href="https://fonts.googleapis.com/icon?family=Material+Icons"
-    />
-
-    <Style />
-    <CssBaseline />
-
-    {/* all this just for a silly date picker */}
-    <MuiPickersUtilsProvider utils={MomentUtils}>
-      {children}
-    </MuiPickersUtilsProvider>
-  </>
-)
-
-export default Root
